cast odometer to float instead of int

+++ custom_components/leafspy/device_tracker.py
@@ -170,7 +170,7 @@ def _parse_see_args(message):
         'attributes': {
             'amp_hours': float(message['AHr']),
             'trip': int(message['Trip']),
-            'odometer': int(message['Odo']),
+            'odometer': float(message['Odo']),
             'battery_temperature': float(message['BatTemp']),
             'outside_temperature': float(message['Amb']),
             'plug_state': PLUG_STATES[int(message['PlugState'])],
